requirements.txt
streamlit
google-genai
Pillow
supabase